// backend/src/types/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  ORGANIZATION_DONOR = 'ORGANIZATION_DONOR',
  INDIVIDUAL_DONOR = 'INDIVIDUAL_DONOR',
  VOLUNTEER = 'VOLUNTEER',
  NGO = 'NGO'
}

export enum DonationStatus {
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  ASSIGNED = 'ASSIGNED',
  PICKED_UP = 'PICKED_UP',
  DELIVERED = 'DELIVERED',
  VERIFIED = 'VERIFIED',
  CANCELLED = 'CANCELLED'
}

export enum VolunteerStatus {
  AVAILABLE = 'AVAILABLE',
  ASSIGNED = 'ASSIGNED',
  ON_THE_WAY = 'ON_THE_WAY',
  ARRIVED_AT_PICKUP = 'ARRIVED_AT_PICKUP',
  FOOD_PICKED_UP = 'FOOD_PICKED_UP',
  DELIVERING = 'DELIVERING',
  DELIVERED = 'DELIVERED'
}